Added model validation to full poll update endpoint
