+++ Shared/DTOs/PollForUpdateDto.cs
@@ -2,12 +2,7 @@
 
 namespace Shared.DTOs
 {
-    public record PollForUpdateDto
+    public record PollForUpdateDto : PollForManipulationDto
     {
-        [Required(ErrorMessage = "The Name of poll is a required field.")]
-        public string? Name { get; init; }
-        public DateTime Deadline { get; init; }
-        public bool IsDisabled { get; init; }
-        public IEnumerable<CandidateForCreationDto>? Candidates { get; init; }
     }
 }
